+++ src/client/client-style.ts
@@ -2,3 +2,5 @@ import type * as d from '../declarations';
 
 export const styles: d.StyleMap = /*@__PURE__*/ new Map();
 export const modeResolutionChain: d.ResolutionHandler[] = [];
+export const setScopedSSR = (_opts: d.HydrateFactoryOptions) => {};
+export const needsScopedSSR = () => false;